add auto-loader for zen_href_link
For a default install of ZC 1.5.5 to then incorporate Ceon URI Mapping,
one of the core files is loaded before the observer auto loader.
As such, that file does not offer the expected uri rewriting.
Two options: either the ZC load sequence is modified to support
the auto loader style and the rewriting (with unknown effect on other plugins
that do not require such a swap, or the guidance within
includes/init_includes/init_observe.php is followed and the necessary
code is used instead of the auto observer system. Ultimately, the problem
is that the core loadPoint is not in the sequence to support full rewrite
capability because the observer for the url function needs to be executed
before the breadcrumbs.

This applies three changes that involve 2 "new" files and basically
modification/removal of an existing file.

This supports loading the observer for monitoring the
function zen_href_link so that the sequence of core files does not
need to be modified. While that may happen independent of this change,
it is no reason to modify or require others to modify their core code.