Move of non-product general
When a product is of product type product_general (product page of product_info), then when selecting to move the product from within the catalog screen works fine because it is product type 1; however, if the product is of a different product type (document general or other product type), then the generated uri for the product is reported as being product_type and the associated handler for that product type.

The above modification uses the current $products_id to determine both the product type and handler for the product being moved as it is not otherwise set at this point in the product move.